@@ -7,7 +7,7 @@ issue_tracker: https://github.com/adaptyteam/AdaptySDK-Flutter/issues
 documentation: https://docs.adapty.io
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
