@@ -1,13 +1,13 @@
 import 'package:adapty_flutter/models/adapty_purchaser_info.dart';
 
 class RestorePurchasesResult {
-  final AdaptyPurchaserInfo purchaserInfo; // nullable
+  final AdaptyPurchaserInfo? purchaserInfo; // nullable
 
   /// iOS specific
-  final String receipt; // nullable
+  final String? receipt; // nullable
 
   // /// Android specific
-  final String googleValidationResult;
+  final String? googleValidationResult;
 
   RestorePurchasesResult.fromJson(Map<String, dynamic> json)
       : purchaserInfo = json.containsKey(_Keys.purchaserInfo) ? AdaptyPurchaserInfo.fromMap(json[_Keys.purchaserInfo]) : null,
