@@ -1,5 +1,5 @@
 extension MapExtension on Map<String, dynamic> {
-  DateTime dateTimeOrNull(String key) {
+  DateTime? dateTimeOrNull(String key) {
     return this.containsKey(key) ? DateTime.tryParse(this[key]) : null;
   }
 }
